Script Push
This script was written to be able to push simple commands to several
Cisco IOS switches or routers.  Use cases would be to create a new vlan
and deploy to multiple switches, change NTP server, change SNMP
information,  or any command that does not need to be dynamic per
device.